requirements.txt
Flask==2.3.3
pymongo==4.5.0
requests==2.31.0
twilio==7.18.0
python-dotenv==1.0.0
